+++ src/main/java/nextstep/courses/domain/Course.java
@@ -1,5 +1,7 @@
 package nextstep.courses.domain;
 
+import nextstep.courses.domain.session.Sessions;
+
 import java.time.LocalDateTime;
 
 public class Course {
