+++ src/main/java/nextstep/courses/domain/session/Image.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.session;
 
 public class Image {
     private String url;
