+++ src/main/java/nextstep/courses/domain/session/ImageFormat.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.session;
 
 import static java.util.Collections.emptySet;
