+++ src/main/java/nextstep/courses/domain/session/ImageSize.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.session;
 
 public class ImageSize {
     private static final int MIN_WIDTH = 300;
