+++ src/test/java/nextstep/courses/domain/session/ImageFormatTest.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.session;
 
 import static org.assertj.core.api.Assertions.assertThat;
